general updates

+++ ASF_Manager/Main.cs
@@ -258,7 +258,7 @@ namespace ASF_Manager
                         
                         string result = Post_Command_Active_Game(bot.BotName, bot.SteamID, Codigo, ID_GAME);
 
-                        if(result != "Timeout")
+                        if(result != "Timeout" && result != "AlreadyPurchased")
                         {
                             File.WriteAllLines(game, arquivo.Skip(1).ToArray());
                         }
@@ -356,6 +356,12 @@ namespace ASF_Manager
                 Log.blue(content);
                 File.AppendAllText("activation_fail.txt", content + "\n");
                 return "DuplicateActivationCode";
+            }
+            else if (content.Contains("Fail/AlreadyPurchased"))
+            {
+                Log.orange(content);
+                Log.pink($"The code will not be discarded, it will be used in the next bot!");
+                return "AlreadyPurchased";
             }
             else
             {
